power element-wise operator reworked (overloading and mixed types)

+++ modules/elementary_functions/src/cpp/Colon.cpp
@@ -114,14 +114,14 @@ real_colon(Class destinationClass, T low, T high, T step)
         return ArrayOf::doubleConstructor(nan(""));
     }
     if (!std::isfinite(low) || !std::isfinite(high) || !std::isfinite(step)) {
-		if (!std::isfinite(low) && !std::isfinite(step)) {
+        if (!std::isfinite(low) && !std::isfinite(step)) {
             if ((low > 0 && step < 0) || (low < 0 && step > 0)) {
                 if (destinationClass == NLS_SINGLE) {
                     return ArrayOf::singleConstructor(nanf(""));
                 }
                 return ArrayOf::doubleConstructor(nan(""));
             }
-		}
+        }
         if (!std::isfinite(low)) {
             if (low > 0) {
                 return ArrayOf::emptyConstructor(Dimensions(1, 0));
@@ -132,9 +132,9 @@ real_colon(Class destinationClass, T low, T high, T step)
                 return ArrayOf::emptyConstructor(Dimensions(1, 0));
             } else {
                 if (destinationClass == NLS_SINGLE) {
-                    return ArrayOf::singleConstructor(low);
+                    return ArrayOf::singleConstructor((single)low);
                 }
-                return ArrayOf::doubleConstructor(low);
+                return ArrayOf::doubleConstructor((double)low);
             }
         }
         if (!std::isfinite(high)) {
